backend/extraction_api/requirements.txt
flask
flask_cors
pymysql
python-dotenv
# logging
jinja2
requests
pdf2image
pytesseract
opencv-contrib-python
pdfplumber
layoutparser
matplotlib 
numpy